toolkit/components/passwordmgr/test/mochitest/test_autocomplete_highlight_non_login.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test form field autofill highlight</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<script>
SimpleTest.requestFlakyTimeout("Giving a chance for the unexpected field highlight to occur");
let readyPromise = registerRunTests();

runInParent(function initLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  let login1 = Cc["@mozilla.org/login-manager/loginInfo;1"]
                        .createInstance(Ci.nsILoginInfo);
  login1.init("http://example.com", "http://autocomplete", null,
              "user1", "pass1", "", "");

  let login2 = Cc["@mozilla.org/login-manager/loginInfo;1"]
                        .createInstance(Ci.nsILoginInfo);
  login2.init("http://example.com", "http://autocomplete", null,
              "user2", "pass2", "", "");

  Services.logins.addLogin(login1);
  Services.logins.addLogin(login2);
});

</script>
<body>
<p id="display"></p>
<div id="content">
  <form id="form1" action="http://autocomplete" onsubmit="return false;">
    <input  type="text"       id="uname">
    <input  type="password"   id="pword">
    <button type="submit">Submit</button>
  </form>
<pre id="test">
<script>
let username = document.getElementById("uname");
let password = document.getElementById("pword");

add_task(async function setup() {
  ok(readyPromise, "check promise is available");
  await readyPromise;

  await SpecialPowers.pushPrefEnv({"set": [["security.insecure_field_warning.contextual.enabled", true]]});
});

add_task(async function test_field_highlight_on_pw_field_autocomplete_insecureWarning() {
  // Press enter on insecure warning and check.
  password.focus();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open popup
  await shownPromise;
  synthesizeKey("KEY_ArrowDown"); // insecure warning
  synthesizeKey("KEY_Enter");

  await new Promise(resolve => setTimeout(resolve, 1000));
  is(document.defaultView.getComputedStyle(password).getPropertyValue("filter"), "none",
     "Highlight is not applied to the password field if enter key is pressed on the insecure warning item");
  is(document.defaultView.getComputedStyle(username).getPropertyValue("filter"), "none",
     "Highlight is not applied to the username field if enter key is pressed on the insecure warning item");

  // Press tab on insecure warning and check.
  password.focus();
  shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open popup
  await shownPromise;
  synthesizeKey("KEY_ArrowDown"); // insecure warning
  synthesizeKey("KEY_Tab");

  await new Promise(resolve => setTimeout(resolve, 1000));
  is(document.defaultView.getComputedStyle(password).getPropertyValue("filter"), "none",
     "Highlight is not applied to the password field if tab key is pressed on the insecure warning item");
  is(document.defaultView.getComputedStyle(username).getPropertyValue("filter"), "none",
     "Highlight is not applied to the username field if tab key is pressed on the insecure warning item");
});

add_task(async function test_field_highlight_on_pw_field_autocomplete_footer() {
  // Press enter on the footer and check.
  password.focus();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open popup
  await shownPromise;
  synthesizeKey("KEY_ArrowUp"); // footer
  synthesizeKey("KEY_Enter");

  await new Promise(resolve => setTimeout(resolve, 1000));
  is(document.defaultView.getComputedStyle(password).getPropertyValue("filter"), "none",
     "Highlight is not applied to the password field if enter key is pressed on the footer item");
  is(document.defaultView.getComputedStyle(username).getPropertyValue("filter"), "none",
     "Highlight is not applied to the username field if enter key is pressed on the footer item");

  runInParent(function cleanUpWindow() {
    const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");
    let window = Services.wm.getMostRecentWindow("navigator:browser");
    window.gBrowser.removeTab(window.gBrowser.selectedTab);
  });

  // Press tab on the footer and check.
  password.focus();
  shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open popup
  await shownPromise;
  synthesizeKey("KEY_ArrowUp"); // footer
  synthesizeKey("KEY_Tab");

  await new Promise(resolve => setTimeout(resolve, 1000));
  is(document.defaultView.getComputedStyle(password).getPropertyValue("filter"), "none",
     "Highlight is not applied to the password field if tab key is pressed on the footer item");
  is(document.defaultView.getComputedStyle(username).getPropertyValue("filter"), "none",
     "Highlight is not applied to the username field if tab key is pressed on the insecure warning item");
});
</script>
</body>
</html>
